+++ pkg/validator/handler.go
@@ -70,13 +70,15 @@ func writeResponse(validator Validator, writer http.ResponseWriter, response *v1
 		Response: response,
 	})
 	if err != nil {
-		validator.Log("level", "error", "message", "unable to serialize response", microerror.JSON(err))
+		validator.Log("level", "error", "message", "unable to serialize response", "stack", microerror.JSON(err))
 		writer.WriteHeader(http.StatusInternalServerError)
 	}
 
 	if _, err := writer.Write(resp); err != nil {
-		validator.Log("level", "error", "message", "unable to write response", microerror.JSON(err))
+		validator.Log("level", "error", "message", "unable to write response", "stack", microerror.JSON(err))
 	}
+
+	validator.Log("level", "info", "message", fmt.Sprintf("Validated request responded with result: %t", response.Allowed))
 }
 
 func errorResponse(uid types.UID, err error) *v1beta1.AdmissionResponse {